2.31
IMPORTANT!! In your Botnak folder:
Delete your TwitchFaces/ folder
Delete your TwitchFaces.txt file

Added:

IRCv3 Support (closes #83 and closes #84)
FrankerFaceZ Support (closes #77 thanks @jbzdarkid)
Reconnect threads (closes #62)
URL-Unshortener, Twitch VOD and YouTube video data info (closes #24)
Emote pop-up menus (try clicking an emote!)

Fixed:

Donation formatting fixed (thanks @dr-kegel & @jbzdarkid)
Potential issue with imgur links (thanks @dr-kegel)
General code cleanup

+++ src/main/java/lib/JSON/XML.java
@@ -358,7 +358,7 @@ public class XML {
                     return Double.valueOf(string);
                 } else if (string.indexOf('e') < 0 && string.indexOf('E') < 0) {
                     Long myLong = new Long(string);
-                    if (myLong.longValue() == myLong.intValue()) {
+                    if (myLong == myLong.intValue()) {
                         return myLong.intValue();
                     } else {
                         return myLong;
